Adjust "skip backwards" behaviour
This changes the "skip backwards" behaviour to be consistent with
the current iOS implementation. Essentially, if the current position
is less than or equal to four seconds from the start of the current
chapter, we jump back to 15 seconds from the end of the previous
chapter (assuming one is available). If the current position is greater
than four, but less than 15 seconds from the start of the chapter,
we jump back to 00:00 in the current chapter.

Testing: All unit tests pass (some tests updated).
         Informal testing with sandbox application.

Affects: https://jira.nypl.org/browse/SIMPLY-1302

+++ org.nypl.audiobook.android.views/src/main/java/org/nypl/audiobook/android/views/PlayerFragment.kt
@@ -92,7 +92,7 @@ class PlayerFragment : android.support.v4.app.Fragment() {
   private lateinit var parameters: PlayerFragmentParameters
 
   private var playerPositionCurrentSpine: PlayerSpineElementType? = null
-  private var playerPositionCurrentOffset: Int = 0
+  private var playerPositionCurrentOffset: Long = 0L
   private var playerEventSubscription: Subscription? = null
   private var playerSleepTimerEventSubscription: Subscription? = null
 
@@ -374,8 +374,8 @@ class PlayerFragment : android.support.v4.app.Fragment() {
     this.log.debug("onProgressBarChanged: {} {}", progress, fromUser)
   }
 
-  private fun hourMinuteSecondTextFromMilliseconds(milliseconds: Int): String {
-    return this.hourMinuteSecondFormatter.print(Duration.millis(milliseconds.toLong()).toPeriod())
+  private fun hourMinuteSecondTextFromMilliseconds(milliseconds: Long): String {
+    return this.hourMinuteSecondFormatter.print(Duration.millis(milliseconds).toPeriod())
   }
 
   private fun hourMinuteSecondTextFromDuration(duration: Duration): String {
@@ -523,7 +523,7 @@ class PlayerFragment : android.support.v4.app.Fragment() {
 
   private fun onEventUpdateTimeRelatedUI(
     spineElement: PlayerSpineElementType,
-    offsetMilliseconds: Int) {
+    offsetMilliseconds: Long) {
 
     this.playerPosition.max = spineElement.duration.standardSeconds.toInt()
     this.playerPosition.isEnabled = true
@@ -533,7 +533,7 @@ class PlayerFragment : android.support.v4.app.Fragment() {
 
     if (!this.playerPositionDragging) {
       this.playerPosition.progress =
-        TimeUnit.MILLISECONDS.toSeconds(offsetMilliseconds.toLong()).toInt()
+        TimeUnit.MILLISECONDS.toSeconds(offsetMilliseconds).toInt()
     }
 
     this.playerTimeMaximum.text =
